// src/shared/types/records.types.ts
export type IndebtedPeople = {
  _id?: string;
  name: string;
  amount: number;
  amountPaid: number;
  isPaid: boolean;
}

export type IndebtedPeopleUI = IndebtedPeople & {
  amountFormatted: string
  amountPaidFormatted: string;
  remainingAmountFormatted: string;
}

export type AccountRecord = {
  _id: string;
  transferRecord?: TransferRecord;
  userId: string;
  shortName: string;
  typeOfRecord: TypeOfRecord;
  description: string;
  amount: number;
  amountFormatted: string;
  date: Date;
  fullDate: string;
  formattedTime: string;
  category: Category | null;
  subCategory: string;
  tag: string[];
  indebtedPeople: IndebtedPeople[];
  account: string;
  budgets: string[];
}

export type ExpenseRecord = AccountRecord & {
  isPaid: boolean;
}

export type IncomeRecord = AccountRecord & {
  expensesPaid: ExpensePaid[];
}

export type ExpensePaid = {
  _id: string;
  shortName: string;
  amount: number;
  amountFormatted: string;
  formattedTime: string;
  fullDate: string;
  isPaid: boolean;
  date: Date;
}

export type DeleteRecordPayload = {
  recordId: string;
}
// ...
export type CreateExpensePayload = {
  account: string;
  amount: number;
  budgets: string[];
  category: string;
  date: Date;
  description: string;
  indebtedPeople: IndebtedPeople[];
  isPaid: boolean;
  linkedBudgets: string[];
  shortName: string;
  subCategory: string;
  tag: string[];
  typeOfRecord: 'expense'
}
export type EditExpensePayload = CreateExpensePayload & {
  recordId: string;
}
export type EditTransferExpensePayload = Omit<CreateExpensePayload, 'typeOfRecord'> & {
  recordId: string;
  typeOfRecord: 'transfer'
}

export type CreateIncomePayload = {
  account: string;
  amount: number;
  budgets: string[];
  category: string;
  date: Date;
  description: string;
  expensesPaid: ExpensePaid[]
  indebtedPeople: IndebtedPeople[];
  shortName: string;
  subCategory: string;
  tag: string[];
  typeOfRecord: 'income'
}

export type EditIncomePayload = CreateIncomePayload & {
  recordId: string;
}
export type EditTransferIncomePayload = Omit<EditIncomePayload, 'typeOfRecord'> & {
  recordId: string;
  typeOfRecord: 'transfer'
}

export type CreateTransferValues = {
  amount: number;
  budgets: string[];
  category: string;
  date: Date;
  description: string;
  shortName: string;
  subCategory: string;
  origin: string;
  destination: string;
  tag: string[];
}

export type TransferExpense = {
  indebtedPeople: never[];
  account: string;
  typeOfRecord: 'transfer';
  isPaid: boolean;
  linkedBudgets: never[];
  amount: number;
  budgets: string[];
  category: string;
  date: Date;
  description: string;
  shortName: string;
  subCategory: string;
  tag: string[];
}

export type TransferIncome = {
  indebtedPeople: never[];
  expensesPaid: ExpensePaid[];
  account: string;
  typeOfRecord: 'transfer';
  amount: number;
  budgets: string[];
  category: string;
  date: Date;
  description: string;
  shortName: string;
  subCategory: string;
  tag: string[];
}

export type CreateTransferPayload = {
  expense: TransferExpense
  income: TransferIncome
}

export interface ExpenseDataResponse {
  data: {
    expense: ExpenseRecord
  }
  error: null;
  message: string[];
  success: boolean;
  version: string;
}

export interface IncomeDataResponse {
  data: {
    income: IncomeRecord
  }
  error: null;
  message: string[];
  success: boolean;
  version: string;
}

export interface TransferDataResponse {
  data: {
    expense: ExpenseRecord
    income: IncomeRecord
  }
  error: null;
  message: string[];
  success: boolean;
  version: string;
}

export interface DeleteExpenseDataResponse {
  data: {
    expense: ExpenseRecord
  }
  error: null;
  message: string[];
  success: boolean;
  version: string;
}

export interface DeleteIncomeDataResponse {
  data: {
    income: IncomeRecord
  }
  error: null;
  message: string[];
  success: boolean;
  version: string;
}